change: task_list_item 분리 -> completed_task_list 중복코드 제거
