+++ lib/modules/home/local_widgets/completed_task_list.dart
@@ -1,10 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:memove_practice/data/task_data.dart';
+import 'package:memove_practice/modules/home/local_widgets/task_list_item.dart';
 import 'package:memove_practice/modules/task_editor/task_editor_page.dart';
 import 'package:memove_practice/theme/theme.dart';
 import 'package:provider/provider.dart';
 
 class CompletedTaskList extends StatelessWidget {
+  final List<Color> textColors;
+  final int currentColorIndex;
+
+  CompletedTaskList({required this.textColors, required this.currentColorIndex});
+
   @override
   Widget build(BuildContext context) {
     final taskData = Provider.of<TaskData>(context);
@@ -12,11 +18,8 @@ class CompletedTaskList extends StatelessWidget {
     return Column(
       children: [
         Container(
-          height: 60,
-          margin: EdgeInsets.only(bottom: 10),
           decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(10),
-            color: getLighterColor(mainColor, 0.85),
+            color: getLighterColor(mainColor, 0.95),
           ),
           alignment: Alignment.center,
           child: Row(
@@ -59,57 +62,7 @@ class CompletedTaskList extends StatelessWidget {
           itemCount: taskData.taskCount,
           itemBuilder: (context, index) {
             if (taskData.taskList[index].isChecked) {
-              return InkWell(
-                onTap: () {
-                  Navigator.push(
-                    context,
-                    PageRouteBuilder(
-                      pageBuilder: (context, animation, secondaryAnimation) => TaskEditPage(taskIndex: index),
-                      transitionsBuilder: (context, animation, secondaryAnimation, child) {
-                        return FadeTransition(
-                          opacity: animation,
-                          child: child,
-                        );
-                      },
-                      transitionDuration: Duration(milliseconds: 150),
-                    ),
-                  );
-                },
-                child: Stack(
-                  children: [
-                    Container(
-                      height: 80,
-                      margin: EdgeInsets.only(bottom: 10),
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(10),
-                        color: getLighterColor(mainColor, 0.85),
-                      ),
-                      alignment: Alignment.center,
-                      child: Text(
-                        taskData.taskList[index].title.content,
-                        style: TextStyle(
-                          fontSize: taskData.taskList[index].title.size,
-                          decoration: taskData.taskList[index].isChecked ? TextDecoration.lineThrough : null,
-                        ),
-                      ),
-                    ),
-                    Align(
-                      alignment: Alignment.centerLeft,
-                      child: SizedBox(
-                        height: 80,
-                        width: 80,
-                        child: IconButton(
-                          onPressed: () {
-                            taskData.changeCheckState(index: index);
-                            taskData.changeIsStrikeThrough(index: index);
-                          },
-                          icon: Icon(taskData.taskList[index].isChecked ? Icons.check : Icons.square_outlined),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              );
+              return TaskListItem(index: index, textColor: textColors[currentColorIndex]);
             } else {
               return Container();
             }
